fixed more warnings
